Refined spec during initial implementation experience.
Added -e option.  Calling programs will want an easy way to tell if
a change occurred, so we offer it via the exit status, but
they should ask for it explicitly, other we
may trigger error detection in other places, when no error has actually
occurred.

Clarified that -o, -g and -m should take effect on FILE even if it's
not replaced.